add project load/save buttons to sing editor
