+++ src/components/Menu/MenuBar/BaseMenuBar.vue
@@ -34,6 +34,8 @@ import TitleBarButtons from "./TitleBarButtons.vue";
 import TitleBarEditorSwitcher from "./TitleBarEditorSwitcher.vue";
 import { useStore } from "@/store";
 import { base64ImageToUri } from "@/helpers/imageHelper";
+import { HotkeyActionType, HotkeyReturnType } from "@/type/preload";
+import { setHotkeyFunctions } from "@/store/setting";
 
 const props =
   defineProps<{
@@ -118,6 +120,65 @@ const openHelpDialog = () => {
     isHelpDialogOpen: true,
   });
 };
+
+const createNewProject = async () => {
+  if (!uiLocked.value) {
+    await store.dispatch("CREATE_NEW_PROJECT", {});
+  }
+};
+
+const saveProject = async () => {
+  if (!uiLocked.value) {
+    await store.dispatch("SAVE_PROJECT_FILE", { overwrite: true });
+  }
+};
+
+const saveProjectAs = async () => {
+  if (!uiLocked.value) {
+    await store.dispatch("SAVE_PROJECT_FILE", {});
+  }
+};
+
+const importProject = () => {
+  if (!uiLocked.value) {
+    store.dispatch("LOAD_PROJECT_FILE", {});
+  }
+};
+
+// 「最近使ったプロジェクト」のメニュー
+const recentProjectsSubMenuData = ref<MenuItemData[]>([]);
+const updateRecentProjects = async () => {
+  const recentlyUsedProjects = await store.dispatch(
+    "GET_RECENTLY_USED_PROJECTS"
+  );
+  recentProjectsSubMenuData.value =
+    recentlyUsedProjects.length === 0
+      ? [
+          {
+            type: "button",
+            label: "最近使ったプロジェクトはありません",
+            onClick: () => {
+              // 何もしない
+            },
+            disabled: true,
+            disableWhenUiLocked: false,
+          },
+        ]
+      : recentlyUsedProjects.map((projectFilePath) => ({
+          type: "button",
+          label: projectFilePath,
+          onClick: () => {
+            store.dispatch("LOAD_PROJECT_FILE", {
+              filePath: projectFilePath,
+            });
+          },
+          disableWhenUiLocked: false,
+        }));
+};
+const projectFilePath = computed(() => store.state.projectFilePath);
+watch(projectFilePath, updateRecentProjects, {
+  immediate: true,
+});
 
 // 「エンジン」メニューのエンジン毎の項目
 const engineSubMenuData = computed<MenuItemData[]>(() => {
@@ -223,7 +284,46 @@ const menudata = computed<MenuItemData[]>(() => [
       closeAllDialog();
     },
     disableWhenUiLocked: false,
-    subMenu: props.fileSubMenuData,
+    subMenu: [
+      ...props.fileSubMenuData,
+      { type: "separator" },
+      {
+        type: "button",
+        label: "新規プロジェクト",
+        onClick: createNewProject,
+        disableWhenUiLocked: true,
+      },
+      {
+        type: "button",
+        label: "プロジェクトを上書き保存",
+        onClick: async () => {
+          await saveProject();
+        },
+        disableWhenUiLocked: true,
+      },
+      {
+        type: "button",
+        label: "プロジェクトを名前を付けて保存",
+        onClick: async () => {
+          await saveProjectAs();
+        },
+        disableWhenUiLocked: true,
+      },
+      {
+        type: "button",
+        label: "プロジェクト読み込み",
+        onClick: () => {
+          importProject();
+        },
+        disableWhenUiLocked: true,
+      },
+      {
+        type: "root",
+        label: "最近使ったプロジェクト",
+        disableWhenUiLocked: true,
+        subMenu: recentProjectsSubMenuData.value,
+      },
+    ],
   },
   {
     type: "root",
@@ -338,6 +438,15 @@ watch(uiLocked, () => {
     subMenuOpenFlags.value = [...Array(menudata.value.length)].map(() => false);
   }
 });
+
+const hotkeyMap = new Map<HotkeyActionType, () => HotkeyReturnType>([
+  ["新規プロジェクト", createNewProject],
+  ["プロジェクトを上書き保存", saveProject],
+  ["プロジェクトを名前を付けて保存", saveProjectAs],
+  ["プロジェクト読み込み", importProject],
+]);
+
+setHotkeyFunctions(hotkeyMap);
 </script>
 
 <style lang="scss">
